sizes and options refactoring for security

+++ src/option/option.service.ts
@@ -4,8 +4,8 @@ import { Repository } from 'typeorm';
 import { Option } from './entity/option.entity';
 import { CreateOptionDto } from './dto/createOption.dto';
 import { Size } from '../size/entity/size.entity';
-
-
+import { Product } from '../product/entity/product.entity';
+import { CurrentUserDto } from '../common/currentUser.dto';
 
 @Injectable()
 export class OptionService {
@@ -15,6 +15,9 @@ export class OptionService {
 
         @InjectRepository(Size)
         private readonly sizeRepository: Repository<Size>,
+
+        @InjectRepository(Product)
+        private readonly productRepository: Repository<Product>,
     ) {}
 
     async create(CreateOptionDto: CreateOptionDto): Promise<Option> {
@@ -65,11 +68,47 @@ export class OptionService {
         }
     }
 
-    async delete(id: string): Promise<void> {
-        const option = await this.optionRepository.findOne({where: {id}});
+    async delete(id: string, user: CurrentUserDto): Promise<void> {
+        const option = await this.optionRepository.findOne({
+            where: {id},
+            relations: ['size', 'size.product']
+        });
         if (!option) {
-          throw new NotFoundException(`Option with id ${id} not found`);
+            throw new NotFoundException(`Option with id ${id} not found`);
         }
+
+        const product = await this.productRepository.findOne({
+            where: {id: option.size.product.id},
+            relations: ['seller']
+        });
+        if(user.userId !== product.seller.id && user.role.name !== 'admin') {
+            throw new ConflictException('You are not authorized to delete this option');
+        }
+
         await this.optionRepository.delete(option.id);
       }
+
+
+    async update(id: string, createOptionDto: CreateOptionDto, user: CurrentUserDto): Promise<void> {
+        const option = await this.optionRepository.findOne({
+            where: {id},
+            relations: ['size', 'size.product']
+        });
+        if (!option) {
+            throw new NotFoundException(`Option with id ${id} not found`);
+        }
+
+        const product = await this.productRepository.findOne({
+            where: {id: option.size.product.id},
+            relations: ['seller']
+        });
+        if(user.userId !== product.seller.id && user.role.name !== 'admin') {
+            throw new ConflictException('You are not authorized to update this option');
+        }
+
+        option.description = createOptionDto.description;
+        option.imageUrl = createOptionDto.imageUrl;
+
+        await this.optionRepository.save(option);
+    }
 }
